\name{testdata}
\alias{testdata}
\docType{data}
\title{a test dataset containing DNA methylation and gene expression measured on BRCA, COAD, and LUAD}
\description{
  100 most varied DNA methylation probes and expressed genes are selected to construct the demo dataset. Additional information to run LRAcluster and labels is also supplied
}
\seealso{
\code{\link{LRAcluster}}
}
